DES-20 Migrate to JUnit 5 !READ COMMIT MSG!
This is a significant change as I needed to change the project organization and dependencies for testing to allow for better refactoring opportunities. You will likely need to run `<path to mvn.cmd> dependency:purge-local-repository in the project directory in order for intellisense to trigger without a build. All tests currently pass except for the ExceptionLinterTest, which is currently serving as a testbed for the new AnalyzerFixture testing architecture.